Improve start up procedure
The sleep before starting the websocket is necessary for the situations
where the application is running and the League Client hasn't been
started. The sleep adds a necessary delay otherwise errors will be
thrown.

Having a separate thread handle user setup helps prevent blocking the
main thread if the application is running for the first time, the League
Client is running but the user hasn't logged in.
The thread also helps keep track of the current user login status

Also implemented an easier way to handle keeping track and closing
the currently running threads.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>